Don't restore token cache, but save context and scope
We parsed 'include' token. We don't need to restore it, because we
provide our own way of handling the entire #include "file.c+". Thus if
we reverted the token back to the parser, we are in a trouble. But we
need to save the Context and Scope RAII, to solve potential issues when
using #include "myFile.C+" after a scope declaration like in the
following example:
void Check(TObject* obj) {
if (obj) cout << "Found the referenced object\n";
else cout << "Error: Could not find the referenced object\n";
}
#include "A.C+"