Add TStreamerOffsets and TStreamerInfo::BuildOffsets().
They will be used as the authoritative interpreter representation of a class.
It will be used for I/O without interpreter definition.
It still needs machinery in rootcling to generate it and in TClass to use it,
including a new TClass state that is somewhere between interpreted (but without
clang::Decl*) and emulated (but with correct offset info).