TabCompletion, more user friendly recovery for interchanging "." and "->".
In ROOT 5 useing var. for a pointer or similarly var-> for an instance of an object would leave the prompt pending but already having in the buffer the already typed input confusing the user. Now it prints the message and autocompletes the prompt with the correct form for the variable in cause.