TTabCom: do not deserialize for classes<TAB> and use gROOT->GetListOfGlobalFunc.
This is a temporary solution to get a reasonable subset of classes. Previously using ".class" and ".namespace" triggered deserialization. We have to implement getting the classes using decls. This uses gClassTable and mapFile. Another issue was with the implementation of TTabCom::GetListOfGlobalFunctions which did not iterate over the MethodInfos. Now using gROOT->GetListOfGlobalFunctions().